requirements.txt
numpy
geopandas
rasterio
xarray
scikit-image
esda
libpysal
pyogrio
rioxarray
python-docx
